Set socket.io 'destroy buffer size' option to Infinity
This is a security feature of Socket.IO which was build to prevent people
from spamming your server with data. But as outlined in
LearnBoost/socket.io#1519 they forgot to clear the internal _dataLength
property. So long running connections or sending frequent data will
result in to random disconnects. Setting this property to infinity solves
the disconnects for now..