Add mechanism to dynamicaly expose NFS shares only to relevant instances
The mechanism is enabled by setting useDynamicNFSShares to true in nefelus.conf.
Notice that the variable filemanager should be set in hub.conf to point to nefelus-s3 endpoint.

The necessary NFS shares are activated by sending a request to the filemanager with the
instance-id, the ip(s) of the instance and the uuids of the mountpoints.
The activate request is sent as soon as the instance is launched and is assigned IP address.
The NFS are deactivated as soon as the instance is up and running and has started communication
with the hub service.

Periodically, a process checks for NFS shares that are activated but the corresponding instance is not
active any more. For such cases, a request is sent to the filemanager to deactivate those shares.