Feat: Support personalised URLs to worker VM consoles
Notes:
a. the activation of the personalized consoles is controled by
   "personaliseConsoleURLs" variable in hub.conf
b. The access to the consoles is controled by a list of users stored in
   /var/run/nefelus/users on the worker VM. The list of users is composed
   by the list of watchers and editors set in the corresponding entry of
   the HUB_REQUEST session
c. a record of the users file is formatted as fields separated by colons
   the fields are : username, access mode (ro/rw), user's full name, user's UUID
d. the worker VM is notified upon change of the collaborators
   of the running session and user consoles are updated accordingly
e. the frontend should retrive the UUID of the user and add to the
   base console URL the parameter: &ui=<UUID>
f. the USER_LOGIN table should be altered by adding a UUID column
g. new vmdashboard, prompt and noVNC consoles and new master should be deployed.